Add slam-evals-smoke task: one RRD per modality for visual inspection
Refactored the synthetic-fixture builder out of tests/conftest.py and into
slam_evals/data/synthetic.py so both pytest and the new tools/smoke.py CLI
share the same generator. The smoke task materialises six tiny (~100 KB,
30 frames, 64×48) RRDs under data/slam-evals/smoke/, one per modality:
mono, mono-vi, stereo, stereo-vi, rgbd, rgbd-vi. Each is openable directly
with `rerun data/slam-evals/smoke/<modality>.rrd` to verify the schema
round-trips through every modality branch.

Wired up as `slam-evals-smoke` in pixi.toml. tests/conftest.py now just
re-exports build_fixture via the shared module.
